+++ kompress-core/src/commonMain/kotlin/dev/karmakrafts/kompress/archiver/ArchiveEntry.kt
@@ -14,28 +14,13 @@
  * limitations under the License.
  */
 
-rootProject.name = "kompress"
+package dev.karmakrafts.kompress.archiver
 
-pluginManagement {
-    repositories {
-        google()
-        mavenCentral()
-        mavenLocal()
-        gradlePluginPortal()
-        maven("https://central.sonatype.com/repository/maven-snapshots")
-    }
+import kotlinx.io.RawSink
+import kotlinx.io.RawSource
+
+interface ArchiveEntry {
+    val offset: Long
+    fun source(): RawSource
+    fun sink(): RawSink
 }
-
-@Suppress("UnstableApiUsage")
-dependencyResolutionManagement {
-    repositories {
-        google()
-        mavenCentral()
-        mavenLocal()
-        maven("https://central.sonatype.com/repository/maven-snapshots")
-    }
-}
-
-enableFeaturePreview("TYPESAFE_PROJECT_ACCESSORS")
-include("kompress-core")
-include("kompress-lz4")