+++ kompress-core/src/commonMain/kotlin/dev/karmakrafts/kompress/archiver/DelegateEntry.kt
@@ -14,28 +14,14 @@
  * limitations under the License.
  */
 
-rootProject.name = "kompress"
+package dev.karmakrafts.kompress.archiver
 
-pluginManagement {
-    repositories {
-        google()
-        mavenCentral()
-        mavenLocal()
-        gradlePluginPortal()
-        maven("https://central.sonatype.com/repository/maven-snapshots")
-    }
+import kotlinx.io.RawSink
+import kotlinx.io.RawSource
+
+open class DelegateEntry(
+    override val offset: Long, private val sourceProvider: () -> RawSource, private val sinkProvider: () -> RawSink
+) : ArchiveEntry {
+    override fun source(): RawSource = sourceProvider()
+    override fun sink(): RawSink = sinkProvider()
 }
-
-@Suppress("UnstableApiUsage")
-dependencyResolutionManagement {
-    repositories {
-        google()
-        mavenCentral()
-        mavenLocal()
-        maven("https://central.sonatype.com/repository/maven-snapshots")
-    }
-}
-
-enableFeaturePreview("TYPESAFE_PROJECT_ACCESSORS")
-include("kompress-core")
-include("kompress-lz4")